\name{paquete_variado-package}
\alias{paquete_variado-package}
\alias{paquete_variado}
\docType{package}
\title{
\packageTitle{paquete_variado}
}
\description{
\packageDescription{paquete_variado}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{paquete_variado}
\packageIndices{paquete_variado}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{paquete_variado}

Maintainer: \packageMaintainer{paquete_variado}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
