\name{paquete_variado}
\alias{paquete_variado-package}
\alias{paquete_variado}
\docType{package}
\title{
\packageTitle{Paquete Variado}
}
\description{
\packageDescription{Contiene tres funciones sencillas}
}
\details{


\packageDESCRIPTION{Contiene tres funciones: hanoi, fibonacci y factorial}
\packageIndices{paquete_variado}
}
\author{
\packageAuthor{paquete_variado}

Maintainer: \packageMaintainer{paquete_variado}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
